feat: isolando o helper
