+++ server/controllers/IndexController.js
@@ -1,3 +1,7 @@
+const fs = require("fs");
+const path = require("path");
+
+
 const controller = {
     index: (req, res) => res.render('index', { title: 'CoLearning' }),
     contact: (req, res) => res.render('contato', { title: 'Contato - CoLearning' }),
